requirements.txt
openai>=1.0.0
numpy
PyPDF2
python-dotenv
python-docx>=0.8.11
fastapi